Add publisher/subscriber to websocket handler to allow external classes to process websocket events

+++ src/main/java/com/cyricc/rpiserver/Broadcaster.java
@@ -13,7 +13,7 @@ public class Broadcaster {
 
     final static Queue<Session> sessions = new ConcurrentLinkedQueue<>();
 
-    static void pushTemp(int temp) {
+    static void pushTemp(double temp) {
         sessions.forEach((session) -> {
             try {
                 session.getRemote().sendString(String.valueOf(temp));
